+++ index.js
@@ -6,5 +6,4 @@ import {AppRegistry} from 'react-native';
 import App from './App';
 import {name as appName} from './app.json';
 
-
 AppRegistry.registerComponent(appName, () => App);
